refactor(tests): unify test strategy — three-layer split with single build-tag gate
Replaces the mixed testing.Short() / build-tag approach with a single
consistent rule: any file requiring a broker carries //go:build integration.

Layer 1 — unit (pkg/kafka/*_test.go)
  No broker, no build tag. go test ./pkg/kafka/... always works.

Layer 2 — watermill compliance (pkg/kafka/*_integration_test.go)
  Verifies the message.Publisher / message.Subscriber interface contract
  using watermill's tests.TestPubSub suite. Requires broker.

Layer 3 — Kafka behaviour (tests/integration/)
  At-least-once semantics, network faults, rebalancing, poison pills.
  Requires broker + Toxiproxy.

Changes:
- Split subscriber_test.go → subscriber_test.go (unit) +
  subscriber_integration_test.go (compliance, //go:build integration)
- Split publisher_test.go → publisher_test.go (unit) +
  publisher_integration_test.go (compliance, //go:build integration)
- Split pubsub_test.go → pubsub_test.go (tombstone comment) +
  pubsub_integration_test.go (compliance, //go:build integration)
- Remove all testing.Short() guards (replaced by build tag)
- Makefile: test-short runs go test ./pkg/kafka/... (no flags needed)
  test-integration runs -tags integration ./... (covers all layers)
- TESTING.md: add three-layer table, file tree, gating rule
