+++ pkg/kafka/publisher_test.go
@@ -1,10 +1,8 @@
 package kafka
 
 import (
-	"context"
 	"errors"
 	"testing"
-	"time"
 
 	"github.com/ThreeDotsLabs/watermill"
 	"github.com/ThreeDotsLabs/watermill/message"
@@ -12,6 +10,9 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/twmb/franz-go/pkg/kgo"
 )
+
+// Unit tests for Publisher — no broker required.
+// Broker-required tests live in publisher_integration_test.go.
 
 func TestNewPublisher_ValidConfig(t *testing.T) {
 	config := DefaultPublisherConfig()
@@ -26,9 +27,7 @@ func TestNewPublisher_ValidConfig(t *testing.T) {
 	assert.NotNil(t, publisher.config.Marshaler)
 	assert.Equal(t, logger, publisher.logger)
 
-	// Clean up
-	err = publisher.Close()
-	assert.NoError(t, err)
+	assert.NoError(t, publisher.Close())
 }
 
 func TestNewPublisher_NilLogger(t *testing.T) {
@@ -41,9 +40,7 @@ func TestNewPublisher_NilLogger(t *testing.T) {
 	assert.NotNil(t, publisher)
 	assert.NotNil(t, publisher.logger)
 
-	// Clean up
-	err = publisher.Close()
-	assert.NoError(t, err)
+	assert.NoError(t, publisher.Close())
 }
 
 func TestNewPublisher_InvalidBrokers(t *testing.T) {
@@ -51,119 +48,35 @@ func TestNewPublisher_InvalidBrokers(t *testing.T) {
 		Brokers: []string{},
 	}
 
-	logger := watermill.NewStdLogger(false, false)
-	publisher, err := NewPublisher(config, logger)
+	publisher, err := NewPublisher(config, watermill.NewStdLogger(false, false))
 
 	assert.Error(t, err)
 	assert.Nil(t, publisher)
 	assert.Contains(t, err.Error(), "brokers must not be empty")
 }
 
-func TestPublisher_Publish_SingleMessage(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration test in short mode")
-	}
-
-	config := DefaultPublisherConfig()
-	config.Brokers = []string{"127.0.0.1:9092"}
-
-	logger := watermill.NewStdLogger(false, false)
-	publisher, err := NewPublisher(config, logger)
-	require.NoError(t, err)
-	defer func() { _ = publisher.Close() }()
-
-	topic := "test-topic-" + watermill.NewUUID()
-
-	// Initialize topic to avoid race conditions with auto-creation
-	sub, err := NewSubscriber(SubscriberConfig{Brokers: config.Brokers}, logger)
-	require.NoError(t, err)
-	err = sub.SubscribeInitialize(topic)
-	require.NoError(t, err)
-	_ = sub.Close()
-
-	msg := message.NewMessage(watermill.NewUUID(), []byte("test payload"))
-	msg.Metadata.Set("key1", "value1")
-
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	msg.SetContext(ctx)
-
-	err = publisher.Publish(topic, msg)
-	require.NoError(t, err)
-}
-
-func TestPublisher_Publish_MultipleMessages(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration test in short mode")
-	}
-
-	config := DefaultPublisherConfig()
-	config.Brokers = []string{"127.0.0.1:9092"}
-
-	logger := watermill.NewStdLogger(false, false)
-	publisher, err := NewPublisher(config, logger)
-	require.NoError(t, err)
-	defer func() { _ = publisher.Close() }()
-
-	topic := "test-topic-" + watermill.NewUUID()
-
-	// Initialize topic to avoid race conditions with auto-creation
-	sub, err := NewSubscriber(SubscriberConfig{Brokers: config.Brokers}, logger)
-	require.NoError(t, err)
-	err = sub.SubscribeInitialize(topic)
-	require.NoError(t, err)
-	_ = sub.Close()
-
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	msgs := []*message.Message{
-		message.NewMessage(watermill.NewUUID(), []byte("payload 1")),
-		message.NewMessage(watermill.NewUUID(), []byte("payload 2")),
-		message.NewMessage(watermill.NewUUID(), []byte("payload 3")),
-	}
-
-	for _, msg := range msgs {
-		msg.SetContext(ctx)
-	}
-
-	err = publisher.Publish(topic, msgs...)
-	require.NoError(t, err)
-}
-
 func TestPublisher_Publish_EmptyMessages(t *testing.T) {
 	config := DefaultPublisherConfig()
 	config.Brokers = []string{"127.0.0.1:9092"}
 
-	logger := watermill.NewStdLogger(false, false)
-	publisher, err := NewPublisher(config, logger)
+	publisher, err := NewPublisher(config, watermill.NewStdLogger(false, false))
 	require.NoError(t, err)
 	defer func() { _ = publisher.Close() }()
 
-	// Publishing zero messages should not error
-	err = publisher.Publish("test-topic")
-	assert.NoError(t, err)
-
-	err = publisher.Publish("test-topic", []*message.Message{}...)
-	assert.NoError(t, err)
+	assert.NoError(t, publisher.Publish("test-topic"))
+	assert.NoError(t, publisher.Publish("test-topic", []*message.Message{}...))
 }
 
 func TestPublisher_Publish_ClosedPublisher(t *testing.T) {
 	config := DefaultPublisherConfig()
 	config.Brokers = []string{"127.0.0.1:9092"}
 
-	logger := watermill.NewStdLogger(false, false)
-	publisher, err := NewPublisher(config, logger)
+	publisher, err := NewPublisher(config, watermill.NewStdLogger(false, false))
 	require.NoError(t, err)
+	require.NoError(t, publisher.Close())
 
-	// Close the publisher
-	err = publisher.Close()
-	require.NoError(t, err)
-
-	// Attempt to publish after close
 	msg := message.NewMessage(watermill.NewUUID(), []byte("test"))
-	topic := "test-topic-" + watermill.NewUUID()
-	err = publisher.Publish(topic, msg)
+	err = publisher.Publish("test-topic-"+watermill.NewUUID(), msg)
 
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "publisher closed")
@@ -173,18 +86,12 @@ func TestPublisher_Close_Idempotent(t *testing.T) {
 	config := DefaultPublisherConfig()
 	config.Brokers = []string{"127.0.0.1:9092"}
 
-	logger := watermill.NewStdLogger(false, false)
-	publisher, err := NewPublisher(config, logger)
+	publisher, err := NewPublisher(config, watermill.NewStdLogger(false, false))
 	require.NoError(t, err)
 
-	// First close
-	err = publisher.Close()
-	assert.NoError(t, err)
+	assert.NoError(t, publisher.Close())
 	assert.True(t, publisher.closed)
-
-	// Second close should not error
-	err = publisher.Close()
-	assert.NoError(t, err)
+	assert.NoError(t, publisher.Close())
 }
 
 func TestPublisher_Publish_WithCustomMarshaler(t *testing.T) {
@@ -192,8 +99,7 @@ func TestPublisher_Publish_WithCustomMarshaler(t *testing.T) {
 	config.Brokers = []string{"127.0.0.1:9092"}
 	config.Marshaler = DefaultMarshaler{}
 
-	logger := watermill.NewStdLogger(false, false)
-	publisher, err := NewPublisher(config, logger)
+	publisher, err := NewPublisher(config, watermill.NewStdLogger(false, false))
 	require.NoError(t, err)
 	defer func() { _ = publisher.Close() }()
 
@@ -205,24 +111,15 @@ func TestPublisher_Publish_MarshalError(t *testing.T) {
 	config.Brokers = []string{"127.0.0.1:9092"}
 	config.Marshaler = &failingMarshaler{}
 
-	logger := watermill.NewStdLogger(false, false)
-	publisher, err := NewPublisher(config, logger)
+	publisher, err := NewPublisher(config, watermill.NewStdLogger(false, false))
 	require.NoError(t, err)
 	defer func() { _ = publisher.Close() }()
 
 	msg := message.NewMessage(watermill.NewUUID(), []byte("test"))
-	topic := "test-topic-" + watermill.NewUUID()
-	err = publisher.Publish(topic, msg)
+	err = publisher.Publish("test-topic-"+watermill.NewUUID(), msg)
 
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "cannot marshal message")
-}
-
-// failingMarshaler is a test marshaler that always returns an error
-type failingMarshaler struct{}
-
-func (f *failingMarshaler) Marshal(topic string, msg *message.Message) (*kgo.Record, error) {
-	return nil, errors.New("marshal failed")
 }
 
 func TestNewPublisher_OTelEnabled_Succeeds(t *testing.T) {
@@ -232,11 +129,17 @@ func TestNewPublisher_OTelEnabled_Succeeds(t *testing.T) {
 	config.Brokers = []string{"127.0.0.1:9092"}
 	config.OTelEnabled = true
 
-	logger := watermill.NewStdLogger(false, false)
-	publisher, err := NewPublisher(config, logger)
+	publisher, err := NewPublisher(config, watermill.NewStdLogger(false, false))
 
 	require.NoError(t, err)
 	assert.NotNil(t, publisher)
 
 	_ = publisher.Close()
+}
+
+// failingMarshaler always returns an error — used to test marshal-error handling.
+type failingMarshaler struct{}
+
+func (f *failingMarshaler) Marshal(_ string, _ *message.Message) (*kgo.Record, error) {
+	return nil, errors.New("marshal failed")
 }
