+++ pkg/kafka/subscriber_test.go
@@ -8,53 +8,8 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestSubscribeInitialize_CreatesTopic(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration test in short mode")
-	}
-
-	config := DefaultSubscriberConfig()
-	config.Brokers = []string{"127.0.0.1:9092"}
-
-	logger := watermill.NewStdLogger(false, false)
-	subscriber, err := NewSubscriber(config, logger)
-	require.NoError(t, err)
-	defer func() { _ = subscriber.Close() }()
-
-	topic := "test-subscribe-initialize-" + watermill.NewUUID()
-
-	// Initialize the topic
-	err = subscriber.SubscribeInitialize(topic)
-	require.NoError(t, err)
-
-	// Verify topic was created by initializing again (should not error)
-	err = subscriber.SubscribeInitialize(topic)
-	require.NoError(t, err)
-}
-
-func TestSubscribeInitialize_TopicAlreadyExists(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration test in short mode")
-	}
-
-	config := DefaultSubscriberConfig()
-	config.Brokers = []string{"127.0.0.1:9092"}
-
-	logger := watermill.NewStdLogger(false, false)
-	subscriber, err := NewSubscriber(config, logger)
-	require.NoError(t, err)
-	defer func() { _ = subscriber.Close() }()
-
-	topic := "test-subscribe-initialize-existing-" + watermill.NewUUID()
-
-	// Create topic first time
-	err = subscriber.SubscribeInitialize(topic)
-	require.NoError(t, err)
-
-	// Second call should not error - topic already exists
-	err = subscriber.SubscribeInitialize(topic)
-	require.NoError(t, err)
-}
+// Unit tests for Subscriber — no broker required.
+// Broker-required tests live in subscriber_integration_test.go.
 
 func TestSubscribeInitialize_ClosedSubscriber(t *testing.T) {
 	config := DefaultSubscriberConfig()
@@ -64,11 +19,9 @@ func TestSubscribeInitialize_ClosedSubscriber(t *testing.T) {
 	subscriber, err := NewSubscriber(config, logger)
 	require.NoError(t, err)
 
-	// Close the subscriber
 	err = subscriber.Close()
 	require.NoError(t, err)
 
-	// Attempt to initialize after close
 	err = subscriber.SubscribeInitialize("test-topic")
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "subscriber closed")
